requirements.txt
django==2.2
pytz==2019.3
django-cleanup==4.0.0
djangorestframework==3.10.3
pyyaml==5.1.2
requests==2.21.0
django-rest-passwordreset==1.0.0